Re-arrange profile and project list page
