+++ src/index.tsx
@@ -7,10 +7,12 @@ import App from "./App";
 import Landing from "./pages/Landing/Landing";
 import PortfolioBricks from "./pages/PortfolioBricks/PortfolioBricks";
 import Profile from "./pages/Profile/Profile";
-import ProjectList from "./pages/ProjectList/ProjectList";
+import FriendList from "./pages/FriendList/FriendList";
+import FavoriteList from "./pages/FavoriteList/FavoriteList";
 import SingleProject from "./pages/SingleProject/SingleProject";
 import CreateNewProject from "./pages/CreateNewProject/CreateNewProject";
 import EditExistProject from "./pages/EditExistProject/EditExistProject";
+import Login from "./pages/Login/Login";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
@@ -22,10 +24,12 @@ root.render(
         <Route index element={<Landing />} />
         <Route path="portfolioBricks" element={<PortfolioBricks />} />
         <Route path="profile" element={<Profile />} />
-        <Route path="projectList" element={<ProjectList />} />
         <Route path="singleProject" element={<SingleProject />} />
         <Route path="createNewProject" element={<CreateNewProject />} />
         <Route path="editExistProject" element={<EditExistProject />} />
+        <Route path="friendList" element={<FriendList />} />
+        <Route path="favoriteList" element={<FavoriteList />} />
+        <Route path="login" element={<Login />} />
         <Route path="*" element={<Navigate to="profile" replace />} />
       </Route>
     </Routes>
