+++ src/pages/Profile/Profile.tsx
@@ -1,7 +1,11 @@
-import { useState, useContext } from "react";
+import { useContext } from "react";
 import styled from "styled-components";
+import { doc, deleteDoc } from "firebase/firestore";
 import ReactLoading from "react-loading";
+import { useNavigate } from "react-router-dom";
 import { useTranslation } from "react-i18next";
+import { db } from "../../context/firebaseSDK";
+import getProjects from "../../utils/getProjects";
 import { AuthContext } from "../../context/authContext";
 
 interface Prop {
@@ -12,6 +16,7 @@ interface Prop {
   focus?: string;
   position?: string;
   buttomLine?: string;
+  img?: string;
 }
 
 const Wrapper = styled.div`
@@ -26,47 +31,22 @@ const Wrapper = styled.div`
 
 const Container = styled.div`
   margin: 0 auto;
-  width: 1200px;
+  width: 80%;
+  max-width: 1500px;
   height: 100%;
   position: relative;
   display: flex;
+  @media screen and (max-width: 1300px) {
+    width: 1200px;
+  }
 `;
 
-const InfoContainer = styled.div`
-  margin: 0 auto;
-  height: 100%;
-  width: 100%;
-  max-width: 460px;
-  position: relative;
-  border: 1px solid #b4b4b4;
-  border-radius: 20px;
-`;
-
-const LoginContainer = styled.div`
+const UserInfoContainer = styled.div`
+  width: 400px;
   padding: 20px;
   display: flex;
   flex-direction: column;
-`;
-
-const SignStatus = styled.button`
-  margin-bottom: 30px;
-  padding: 0;
-  width: 50%;
-  height: 60px;
-  font-size: 20px;
-  color: #313538;
-  border: none;
-  border-bottom: ${(props: Prop) => props.buttomLine || "none"};
-  border-radius: ${(props: Prop) => props.position || "0px 20px 0px 0px"};
-  background-color: ${(props) => props.color || "#fff"};
-  & + & {
-    border-left: 1px solid #b4b4b4;
-  }
-  @media screen and (max-width: 1279px) {
-    margin-bottom: 24px;
-    height: 40px;
-    font-size: 16px;
-  }
+  align-items: center;
 `;
 
 const Avatar = styled.div`
@@ -77,7 +57,7 @@ const Avatar = styled.div`
   background-position: center;
 `;
 
-const Context = styled.div`
+const UserInfo = styled.div`
   margin-top: 20px;
   font-size: 24px;
   color: #3c3c3c;
@@ -87,73 +67,69 @@ const Context = styled.div`
   }
 `;
 
-const Input = styled.input`
-  padding: 6px 10px;
-  width: 320px;
-  height: 50px;
-  color: #3c3c3c;
-  font-size: 18px;
-  background-color: #f0f0f090;
-  border: 1px solid gray;
-  & + & {
-    margin-top: 30px;
-  }
-  &:focus {
-    outline: none;
-    background-color: #61616130;
-  }
+const ProjectListContainer = styled.div`
+  margin: 0 auto;
+  width: 1200px;
+  height: 100%;
+  position: relative;
+  display: flex;
+  flex-direction: column;
 `;
 
-const SignContainer = styled.div`
+const ProjectHeaderContainer = styled.div`
+  padding-bottom: 20px;
   display: flex;
 `;
 
-const SignBtn = styled.button`
-  margin-top: 30px;
-  padding: 0 20px;
-  height: 40px;
-  color: #3c3c3c;
-  font-size: 18px;
-  background-color: transparent;
-  border: 1px solid #3c3c3c;
-  &:hover {
-    box-shadow: 1px 1px 5px #616161;
-  }
-  &:focus {
-    outline: none;
-  }
+const ProjectsContainer = styled.div`
+  display: flex;
+  flex-direction: column;
 `;
 
-const LoginOptionsContainer = styled.div`
-  margin: 0 auto;
-  width: 85%;
-  max-width: 380px;
-  height: 40px;
-  position: relative;
-`;
-
-const LoginOptionsText = styled.div`
-  height: 40px;
-  padding: 0 20px;
-  background-color: #ffffff;
-  color: #787878;
-  position: absolute;
-  left: 50%;
-  transform: translateX(-50%);
-  @media screen and (max-width: 1279px) {
-    height: 30px;
-    padding: 0 10px;
-    font-size: 14px;
-  }
-`;
-
-const LoginOptionsLine = styled.div`
+const SingleProjectContainer = styled.div`
+  height: 200px;
   width: 100%;
-  height: 10px;
-  border-bottom: 1px solid #787878;
-  @media screen and (max-width: 1279px) {
-    height: 8px;
+  display: flex;
+  border: 1px solid black;
+  & + & {
+    margin-top: 20px;
   }
+`;
+
+const ProjectLeftContainer = styled.div`
+  padding: 10px;
+  height: 200px;
+  width: 40%;
+  display: flex;
+  flex-direction: column;
+`;
+
+const ProjectTitle = styled.div`
+  font-size: 24px;
+`;
+
+const ProjectRightContainer = styled.div`
+  height: 200px;
+  width: 60%;
+  display: flex;
+  align-items: center;
+`;
+
+const PhotoUrl = styled.div`
+  width: 180px;
+  height: 180px;
+  background-image: ${(props: Prop) => props.img};
+  background-position: center;
+  background-size: cover;
+  & + & {
+    margin-left: 10px;
+  }
+`;
+
+const Button = styled.button`
+  margin-top: 10px;
+  width: 170px;
+  height: 40px;
 `;
 
 const Loading = styled(ReactLoading)`
@@ -162,70 +138,35 @@ const Loading = styled(ReactLoading)`
 
 function Profile() {
   const { t } = useTranslation();
+  const navigate = useNavigate();
   const {
     isLogin,
     isLoading,
     name,
     email,
     avatar,
-    signUp,
-    emailSignInHandler,
-    googleLoginHandler,
-    facebookLoginHandler,
-    logout,
+    userId,
+    userProjects,
+    setSingleProjectId,
+    setUserProjects,
   } = useContext(AuthContext);
-  const [inputName, setInputName] = useState("");
-  const [inputEmail, setInputEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
-  const [signin, setSignin] = useState(true);
 
-  function signInHandler() {
-    if (!inputEmail || !password) {
-      alert(t("email_and_password_input_check"));
-      return;
-    }
-    if (
-      inputEmail.search(
-        /^\w+((-\w+)|(\.\w+))*@[A-Za-z0-9]+((\.|-)[A-Za-z0-9]+)*\.[A-Za-z]+$/
-      )
-    ) {
-      alert(t("email_format_check"));
-      return;
-    }
-    if (password.length < 8) {
-      alert(t("password_letter_check"));
-      return;
-    }
-    emailSignInHandler(inputEmail, password);
+  function toSingleProjectPage(projectId: string) {
+    setSingleProjectId(projectId);
+    navigate("/singleProject");
   }
 
-  function signUpHandler() {
-    if (!inputName) {
-      alert(t("name_input_check"));
-      return;
-    }
-    if (!inputEmail || !password) {
-      alert(t("email_and_password_input_check"));
-      return;
-    }
-    if (
-      inputEmail.search(
-        /^\w+((-\w+)|(\.\w+))*@[A-Za-z0-9]+((\.|-)[A-Za-z0-9]+)*\.[A-Za-z]+$/
-      )
-    ) {
-      alert(t("email_format_check"));
-      return;
-    }
-    if (password.length < 8) {
-      alert(t("password_letter_check"));
-      return;
-    }
-    if (password !== confirmPassword) {
-      alert(t("password_confirm_failed"));
-      return;
-    }
-    signUp(inputEmail, password, inputName);
+  function toEditExistProjectPage(projectId: string) {
+    setSingleProjectId(projectId);
+    navigate("/editExistProject");
+  }
+
+  async function deleteProjectHandler(projectId: string) {
+    const ans = window.confirm(t("delete_project_warning"));
+    if (ans === false) return;
+    await deleteDoc(doc(db, "projects", projectId));
+    const userProjectsData = await getProjects(userId);
+    setUserProjects(userProjectsData);
   }
 
   if (isLoading) {
@@ -236,99 +177,65 @@ function Profile() {
     );
   }
 
+  if (!isLogin) {
+    alert("please login first");
+    navigate("/login");
+  }
+
   return (
     <Wrapper>
       <Container>
-        {isLogin ? (
-          <LoginContainer>
-            <Avatar url={`url(${avatar})`} />
-            <Context size="24px">{name}</Context>
-            <Context size="20px">{email}</Context>
-            <SignBtn onClick={logout}>{t("logout")}</SignBtn>
-          </LoginContainer>
-        ) : (
-          <InfoContainer>
-            <SignStatus
-              onClick={() => setSignin(true)}
-              color={signin ? "" : "#f0f0f0"}
-              position="20px 0px 0px 0px"
-              buttomLine={signin ? "" : "1px solid #b4b4b4"}
-            >
-              {t("login")}
-            </SignStatus>
-            <SignStatus
-              onClick={() => setSignin(false)}
-              color={signin ? "#f0f0f0" : ""}
-              buttomLine={signin ? "1px solid #b4b4b4" : ""}
-            >
-              {t("sign_up")}
-            </SignStatus>
-            {signin ? (
-              <LoginContainer>
-                <Input
-                  placeholder={t("input_email")}
-                  onChange={(e) => setInputEmail(e.target.value)}
-                />
-                <Input
-                  placeholder={t("input_password")}
-                  type="password"
-                  onChange={(e) => setPassword(e.target.value)}
-                />
-                <SignContainer>
-                  <SignBtn onClick={() => signInHandler()}>
-                    {t("login")}
-                  </SignBtn>
-                </SignContainer>
-              </LoginContainer>
-            ) : (
-              <LoginContainer>
-                <Input
-                  placeholder={t("input_name")}
-                  onChange={(e) => setInputName(e.target.value)}
-                />
-                <Input
-                  placeholder={t("input_email")}
-                  onChange={(e) => setInputEmail(e.target.value)}
-                />
-                <Input
-                  placeholder={t("input_password")}
-                  type="password"
-                  onChange={(e) => setPassword(e.target.value)}
-                />
-                <Input
-                  placeholder={t("check_input_password")}
-                  type="password"
-                  onChange={(e) => setConfirmPassword(e.target.value)}
-                />
-                <SignContainer>
-                  <SignBtn onClick={() => signUpHandler()}>
-                    {t("sign_up")}
-                  </SignBtn>
-                </SignContainer>
-              </LoginContainer>
-            )}
-            <LoginOptionsContainer>
-              <LoginOptionsText>or</LoginOptionsText>
-              <LoginOptionsLine />
-            </LoginOptionsContainer>
-            <LoginContainer>
-              <SignBtn
-                onClick={() => {
-                  googleLoginHandler();
-                }}
-              >
-                {t("continue_with_google")}
-              </SignBtn>
-              <SignBtn
-                onClick={() => {
-                  facebookLoginHandler();
-                }}
-              >
-                {t("continue_with_facebook")}
-              </SignBtn>
-            </LoginContainer>
-          </InfoContainer>
-        )}
+        <UserInfoContainer>
+          <Avatar url={`url(${avatar})`} />
+          <UserInfo size="24px">{name}</UserInfo>
+          <UserInfo size="20px">{email}</UserInfo>
+          <UserInfo size="20px">Introduction</UserInfo>
+          <UserInfo size="18px">Hello, I am orange!</UserInfo>
+        </UserInfoContainer>
+        <ProjectListContainer>
+          <ProjectHeaderContainer>
+            <ProjectTitle>{t("project_list")}</ProjectTitle>
+          </ProjectHeaderContainer>
+          {userProjects.length === 0 ? (
+            ""
+          ) : (
+            <ProjectsContainer>
+              {userProjects.map((projectData) => (
+                <SingleProjectContainer key={projectData.projectId}>
+                  <ProjectLeftContainer>
+                    <ProjectTitle>{projectData.title}</ProjectTitle>
+                    {/* <Text>{projectData.time.toDate()}</Text> */}
+                    <Button
+                      onClick={() => toSingleProjectPage(projectData.projectId)}
+                    >
+                      {t("view_project_detail")}
+                    </Button>
+                    <Button
+                      onClick={() =>
+                        toEditExistProjectPage(projectData.projectId)
+                      }
+                    >
+                      {t("edit_again")}
+                    </Button>
+                    <Button
+                      onClick={() =>
+                        deleteProjectHandler(projectData.projectId)
+                      }
+                    >
+                      {t("delete_project")}
+                    </Button>
+                  </ProjectLeftContainer>
+                  <ProjectRightContainer>
+                    {projectData.pages[0].url &&
+                      projectData.pages[0].url.map((singleUrl: string) => (
+                        <PhotoUrl key={singleUrl} img={`url(${singleUrl})`} />
+                      ))}
+                  </ProjectRightContainer>
+                </SingleProjectContainer>
+              ))}
+            </ProjectsContainer>
+          )}
+        </ProjectListContainer>
       </Container>
     </Wrapper>
   );
